refactoring to use switch/cases
