+++ config.json
@@ -1,0 +1,5 @@
+{
+	"prefix": "!",
+	"token": "NjAwMDk2NjM0NTU4MjE4MzE1.XpKBSQ.MHIDfVtSbwQRCWI5ofkT7N0aivA",
+	"giphyKey": "CdtKfRQXmimfqoRUNgxG8ZgJCxc7IlfY"
+}
