+++ giphy.js
@@ -1,10 +1,10 @@
-const auth = require("./auth.json")
+const {giphyKey} = require("./config.json")
 const fetch = require('node-fetch')
-const tags = require('./tag.js')
+const {funTag, surpriseTag} = require('./tag.js')
 
 const giphyManager = {
     surprise: (message) => {
-        return fetch(`http://api.giphy.com/v1/gifs/random?tag=${tags.surpriseTag}&lang=en&api_key=${auth.giphyKey}`, {
+        return fetch(`http://api.giphy.com/v1/gifs/random?tag=${surpriseTag}&lang=en&api_key=${giphyKey}`, {
             method: "GET",
             headers: {
                 'Content-Type': 'application/json'
@@ -16,7 +16,7 @@ const giphyManager = {
             })
     },
     fun: (message) => {
-        return fetch(`http://api.giphy.com/v1/gifs/random?tag=${tags.funTag}&lang=en&api_key=${auth.giphyKey}`, {
+        return fetch(`http://api.giphy.com/v1/gifs/random?tag=${funTag}&lang=en&api_key=${giphyKey}`, {
             method: "GET",
             headers: {
                 'Content-Type': 'application/json'
