+++ package.json
@@ -1,13 +1,14 @@
 {
-    "name": "greeter-bot",
+    "name": "clarence",
     "version": "1.0.0",
-    "description": "My First Discord Bot",
-    "main": "bot.js",
-    "author": "Your Name",
+    "description": "Clarence mutes and insults people, what more would you want?",
+    "main": "discord-bot.js",
+    "scripts": {
+        "start": "node discord-bot.js"
+    },
+    "author": "Madi Peper",
     "dependencies": {
-        "discord.io": "https://github.com/woor/discord.io/tarball/gateway_v6",
         "discord.js": "^11.5.1",
-        "node-fetch": "^2.6.1",
-        "winston": "^3.2.1"
+        "node-fetch": "^2.6.1"
     }
 }
